Support surface meshes in batched LOR assembly
